@@ -15,50 +15,54 @@
   },
   "homepage": "https://github.com/stackstorm/st2flow",
   "devDependencies": {
-    "autoprefixer-core": "5.2.0",
-    "babel-eslint": "5.0.0",
-    "babelify": "6.1.2",
-    "browserify": "10.2.3",
-    "chai": "3.0.0",
-    "del": "1.2.0",
-    "eslint": "0.24.1",
-    "eslint-plugin-react": "3.1.0",
-    "git-rev-sync": "1.4.0",
-    "gulp": "3.9.0",
+    "autoprefixer": "6.0.1",
+    "babel-eslint": "6.1.2",
+    "babel-plugin-transform-class-properties": "^6.11.5",
+    "babel-plugin-transform-object-rest-spread": "^6.8.0",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "babelify": "7.3.0",
+    "browserify": "13.1.0",
+    "chai": "3.5.0",
+    "del": "2.2.2",
+    "eslint": "3.5.0",
+    "eslint-plugin-react": "6.2.2",
+    "git-rev-sync": "1.6.0",
+    "gulp": "3.9.1",
     "gulp-cached": "1.1.0",
-    "gulp-eslint": "0.15.0",
-    "gulp-fontello": "0.2.0",
-    "gulp-header": "1.7.1",
-    "gulp-jshint": "1.11.0",
-    "gulp-mocha": "2.1.0",
-    "gulp-plumber": "1.0.1",
-    "gulp-postcss": "5.1.8",
-    "gulp-size": "1.2.1",
-    "gulp-sourcemaps": "1.5.2",
-    "gulp-util": "3.0.5",
+    "gulp-eslint": "3.0.1",
+    "gulp-fontello": "0.4.6",
+    "gulp-header": "1.8.8",
+    "gulp-jshint": "2.0.1",
+    "gulp-mocha": "3.0.1",
+    "gulp-plumber": "1.1.0",
+    "gulp-postcss": "6.2.0",
+    "gulp-size": "2.1.0",
+    "gulp-sourcemaps": "1.6.0",
+    "gulp-util": "3.0.7",
     "gulp-webserver": "0.9.1",
-    "postcss-color-function": "1.3.2",
-    "postcss-import": "5.2.2",
-    "postcss-mq-keyframes": "0.2.1",
-    "postcss-nested": "0.3.2",
-    "postcss-custom-properties": "4.1.0",
+    "postcss-color-function": "2.0.1",
+    "postcss-custom-properties": "5.0.1",
+    "postcss-import": "8.1.2",
+    "postcss-mq-keyframes": "0.3.0",
+    "postcss-nested": "1.0.0",
     "vinyl-buffer": "1.0.0",
     "vinyl-source-stream": "1.1.0",
-    "watchify": "3.3.0"
+    "watchify": "3.7.0"
   },
   "dependencies": {
-    "URIjs": "1.16.0",
+    "URIjs": "1.16.1",
     "brace": "0.8.0",
-    "dagre": "0.7.2",
+    "dagre": "0.7.4",
     "director": "1.2.8",
-    "lodash": "3.9.3",
+    "lodash": "4.15.0",
     "mixin": "0.2.0",
-    "moment": "2.10.6",
-    "normalize.css": "3.0.3",
-    "object.observe": "0.2.4",
-    "react": "15.0.1",
-    "react-dom": "15.0.1",
-    "react-time": "4.0.2",
+    "moment": "2.15.0",
+    "normalize.css": "4.2.0",
+    "object.observe": "0.2.6",
+    "react": "15.3.1",
+    "react-dom": "15.3.1",
+    "react-time": "4.2.0",
     "revalidator": "0.3.1",
     "st2client": "1.0.0"
   }
