@@ -1,26 +1,23 @@
 import _ from 'lodash';
 import React from 'react';
-import { EventEmitter } from 'events';
 
-import Arrow from './util/arrow';
-import bem from './util/bem';
-import Vector from './util/vector';
-import { pack, unpack } from './util/packer';
+import Arrow from '../util/arrow';
+import bem from '../util/bem';
+import Vector from '../util/vector';
+import { pack, unpack } from '../util/packer';
 
 const st2Class = bem('viewer')
     ;
 
-import Node from './canvas/node';
-import Edge from './canvas/edge';
-import Label from './canvas/label';
+import Node from './node';
+import Edge from './edge';
+import Label from './label';
 
 let hiddenNode;
 
-class CanvasController extends EventEmitter {}
-
 export default class Canvas extends React.Component {
   static propTypes = {
-    graph: React.PropTypes.object
+    model: React.PropTypes.object
   }
 
   state = {
@@ -59,15 +56,37 @@ export default class Canvas extends React.Component {
   }
 
   componentDidMount() {
-    this._canvas = new CanvasController();
+    this.props.model.on('parse', () => {
+      this.draw();
 
-    this._canvas.on('select', (selected) => {
+      // If at least one node doesn't have coordinates, relayout them all, but
+      // not embed coordinates just yet.
+      const nodes = this.props.model.nodes();
+      if (!_.isEmpty(nodes)) {
+        const hasCoords = (name) => {
+          const { x, y } = this.props.model.node(name);
+
+          return x !== undefined && y !== undefined;
+        };
+
+        if (!_.some(nodes, hasCoords)) {
+          this.props.model.graph.layout();
+        }
+      }
+
+      this.show(this.props.model.selected);
+    });
+
+    this.props.model.on('select', (selected) => {
       this.setState({ selected });
     });
   }
 
-  on(...args) {
-    return this._canvas.on(...args);
+  componentDidUpdate() {
+    if (this._pendingEdit) {
+      this.edit(this._pendingEdit);
+      this._pendingEdit = void 0;
+    }
   }
 
   edit(name) {
@@ -79,17 +98,22 @@ export default class Canvas extends React.Component {
   }
 
   show(name) {
-    const node = this._canvas.graph.node(name)
-        , view = this.refs.viewer
-        , { x, y, width, height } = node
+    const node = this.props.model.node(name)
+        , view = this.refs.viewer;
+
+    if (!node) {
+      return;
+    }
+
+    const { x, y, width, height } = node
         , { scrollLeft, scrollTop, clientWidth, clientHeight } = view
-        , [ viewWidth, viewHeight ] = [clientWidth, clientHeight]//this.toInner(clientWidth, clientHeight)
+        , [ viewWidth, viewHeight ] = [clientWidth, clientHeight]
         ;
 
     const A = new Vector(scrollLeft, scrollTop)
         , B = new Vector(x, y)
         , C = new Vector(x + width, y + height)
-        , D = new Vector(...[scrollLeft + viewWidth, scrollTop + viewHeight])//this.toInner(scrollLeft + viewWidth, scrollTop + viewHeight))
+        , D = new Vector(...[scrollLeft + viewWidth, scrollTop + viewHeight])
         , AB = B.subtract(A)
         , CD = D.subtract(C)
         ;
@@ -111,7 +135,8 @@ export default class Canvas extends React.Component {
     }
   }
 
-  draw(graph) {
+  draw() {
+    const graph = this.props.model.graph;
     const nodes = graph._nodes;
     const edges = _.mapValues(graph._edgeObjs, (edge, key) => {
       const { v, w } = edge;
@@ -124,17 +149,42 @@ export default class Canvas extends React.Component {
     });
 
     this.setState({ nodes, edges });
-    this._canvas.graph = graph;
   }
 
   handleLabelClick(e, edge) {
     const { v, w } = edge;
-    this._canvas.emit('disconnect', { v: v.name, w: w.name });
+    this.props.model.disconnect(v.name, w.name);
   }
 
-  handleNodeSelect(e, name) {
-    this._canvas.emit('select', name, e);
-    // this.setState({ selected });
+  handleNodeSelect(event, name) {
+    const SHIFT = 1
+        , ALT = 2
+        , CTRL = 4
+        , META = 8
+        ;
+
+    const mode =
+      event.shiftKey * SHIFT +
+      event.altKey * ALT +
+      event.ctrlKey * CTRL +
+      event.metaKey * META;
+
+    switch(mode) {
+      case SHIFT:
+        this.props.model.connect(this.props.model.selected, name, 'success');
+        break;
+      case ALT:
+        this.props.model.connect(this.props.model.selected, name, 'error');
+        break;
+      case SHIFT + ALT:
+        this.props.model.connect(this.props.model.selected, name, 'complete');
+        break;
+      case META:
+        this.edit(name);
+        break;
+      default:
+        this.props.model.select(name);
+    }
   }
 
   handleNodePick(e, name) {
@@ -168,15 +218,20 @@ export default class Canvas extends React.Component {
   }
 
   handleNodeRename(e, name, value) {
-    this._canvas.emit('rename', name, value);
+    return this.props.model.rename(name, value);
   }
 
   handleNodeDelete(e, name) {
-    this._canvas.emit('delete', name);
+    this.props.model.delete(name);
+    this.focus();
   }
 
   handleNodeConnect(e, source, destination, type) {
-    this._canvas.emit('link', source, destination, type);
+    if (type) {
+      this.props.model.connect(source, destination, type);
+    } else {
+      this.props.model.disconnect(source, destination, type);
+    }
   }
 
   handleCanvasDragEnter(e) {
@@ -206,12 +261,12 @@ export default class Canvas extends React.Component {
     packet = e.dataTransfer.getData('nodePack');
     if (packet) {
       let { name, offsetX, offsetY } = unpack(packet)
-        , {offsetX: x, offsetY: y} = e.nativeEvent // Relative to itself (Viewer)
+        , {offsetX: x, offsetY: y} = e.nativeEvent
         ;
 
-      [x, y] = [x - offsetX, y - offsetY]; //this.toInner(x - offsetX, y - offsetY);
+      [x, y] = [x - offsetX, y - offsetY];
 
-      this._canvas.emit('move', name, x, y);
+      this.props.model.move(name, x, y);
       this.setState({ active: false });
       return;
     }
@@ -219,12 +274,12 @@ export default class Canvas extends React.Component {
     packet = e.dataTransfer.getData('actionPack');
     if (packet) {
       let { action } = unpack(packet)
-        , {offsetX: x, offsetY: y} = e.nativeEvent // Relative to itself (Viewer)
+        , {offsetX: x, offsetY: y} = e.nativeEvent
         ;
 
-      // [x, y] = this.toInner(x, y);
+      const node = this.props.model.create(action.ref, x, y);
+      this._pendingEdit = node.name;
 
-      this._canvas.emit('create', action, x, y);
       this.setState({ active: false });
       return;
     }
@@ -295,6 +350,30 @@ export default class Canvas extends React.Component {
     this.setState({ scale });
   }
 
+  handleKeyDown(e) {
+    const BACKSPACE = 8
+        , DELETE = 46
+        , Z = 90
+        ;
+
+    switch(e.keyCode) {
+      case BACKSPACE:
+      case DELETE:
+        e.preventDefault();
+        this.props.model.delete(this.props.model.selected);
+        break;
+      case Z:
+        if (!e.ctrlKey && !e.metaKey) {
+          return;
+        }
+        if (e.shiftKey) {
+          this.props.model.redo();
+        } else {
+          this.props.model.undo();
+        }
+    }
+  }
+
   getSvgSize() {
     const { nodes={}, scale } = this.state;
 
@@ -336,8 +415,9 @@ export default class Canvas extends React.Component {
 
     const viewerProps = {
       className: st2Class(null),
-      tabindex: -1,
-      ref: 'viewer'
+      tabIndex: '-1',
+      ref: 'viewer',
+      onKeyDown: (e) => this.handleKeyDown(e)
     };
 
     if (this.state.active) {
