@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
-import Sector from './models/sector';
+import Sector from '../model/sector';
 
-export default class Definition {
+export default class BaseDefinition {
   handler(type, spec, parser) {
     return (line, lineNum, state) => {
       const task = state.currentTask;
