@@ -1,10 +1,10 @@
 import _ from 'lodash';
 
-import Definition from '../definition';
-import Sector from '../models/sector';
-import Task from '../models/task';
+import BaseDefinition from './base';
+import Sector from '../model/sector';
+import Task from '../model/task';
 
-export default class ChainDefinition extends Definition {
+export default class ChainDefinition extends BaseDefinition {
   constructor (model) {
     super();
 
