@@ -1,10 +1,10 @@
 import _ from 'lodash';
 
-import Definition from '../definition';
-import Sector from '../models/sector';
-import Workflow from '../models/workflow';
+import BaseDefinition from './base';
+import Sector from '../model/sector';
+import Workflow from '../model/workflow';
 
-export default class MistralDefinition extends Definition {
+export default class MistralDefinition extends BaseDefinition {
   constructor (model) {
     super();
 
