@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import Range from './util/range';
+import Range from '../util/range';
 
 
 export class YaqlCompletion {
@@ -96,7 +96,7 @@ export class InputCompletion {
   actions = {}
 
   update(actions) {
-    const keys = _.pluck(actions, 'ref');
+    const keys = _.map(actions, 'ref');
     const values = _.map(actions, (action) => {
       return _.map(action.parameters, (prop, name) => {
         return _.assign({}, prop, {name});
