@@ -3,7 +3,7 @@ import dagre from 'dagre';
 import { EventEmitter } from 'events';
 import mixin from 'mixin';
 
-import Node from './models/node';
+import Node from './node';
 
 export default class Graph extends mixin(dagre.graphlib.Graph, EventEmitter) {
   constructor() {
@@ -57,17 +57,6 @@ export default class Graph extends mixin(dagre.graphlib.Graph, EventEmitter) {
     });
 
     _.each(nodes, v => this.removeNode(v));
-  }
-
-  select(name) {
-    this.__selected__ = name;
-    this.emit('select', name);
-  }
-  get selected() {
-    return this.node(this.__selected__);
-  }
-  isSelected(name) {
-    return this.__selected__ === name;
   }
 
   connect(source, targets, type='success') {
