@@ -1,6 +1,6 @@
 import React from 'react';
 
-import bem from './util/bem';
+import bem from '../util/bem';
 import Control from './control';
 
 const st2Class = bem('controls')
