@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import React from 'react';
 
-import api from './api';
-import bem from './util/bem';
+import api from '../api';
+import bem from '../util/bem';
 import Control from './control';
 import TooltipItem from './tooltipitem';
 
