@@ -1,9 +1,9 @@
 import React from 'react';
 
-import bem from './util/bem';
+import bem from '../util/bem';
 
-import settings from '../lib/settings';
-import Handle from './canvas/handle';
+import settings from '../settings';
+import Handle from '../canvas/handle';
 
 const st2Class = bem('guide')
     ;
