@@ -349,7 +349,7 @@ export default class Meta extends React.Component {
             }
           }
         },
-        options: _(this.state.packs).pluck('name').union(['default']).value()
+        options: _(this.state.packs).map('name').union(['default']).value()
       }, {
         name: 'dot',
         type: 'comment',
