@@ -1,12 +1,12 @@
 import _ from 'lodash';
 import React from 'react';
 
-import ACTIONS from './util/default-actions';
-import bem from './util/bem';
-import { pack } from './util/packer';
+import ACTIONS from '../util/default-actions';
+import bem from '../util/bem';
+import { pack } from '../util/packer';
 
 
-import icons from './util/icon';
+import icons from '../util/icon';
 
 const st2Class = bem('palette')
     ;
