@@ -1,7 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
-import bem from './util/bem';
+import bem from '../util/bem';
 
 const st2Class = bem('panel')
     ;
@@ -28,7 +27,7 @@ export default class Panel extends React.Component {
   }
 
   componentDidMount() {
-    this.setState({ size: ReactDOM.findDOMNode(this).clientWidth });
+    this.setState({ size: this.refs.self.clientWidth });
     document.addEventListener('mouseup', this.handleMouseUp);
     document.addEventListener('mousemove', this.handleMouseMove);
   }
@@ -71,7 +70,8 @@ export default class Panel extends React.Component {
 
   render() {
     const props = {
-      className: st2Class(null)
+      className: st2Class(null),
+      ref: 'self'
     }
     ;
 
