@@ -2,8 +2,8 @@ import moment from 'moment';
 import React from 'react';
 import Time from 'react-time';
 
-import api from './api';
-import bem from './util/bem';
+import api from '../api';
+import bem from '../util/bem';
 
 const st2Class = bem('controls')
     ;
