@@ -1,3 +1,1 @@
-module.exports = require('babelify/node_modules/babel-core/register')({
-  optional: ['es7.classProperties']
-});
+module.exports = require('babelify/node_modules/babel-core/register')();
